baseComponentsMoved
